refactor and debug
