+++ backend/api/views.py
@@ -8,7 +8,7 @@ from django.shortcuts import get_object_or_404
 from django.http import JsonResponse
 from django.conf import settings
 from abc import ABC, abstractmethod
-from api.serializer import *
+from api import serializers
 from events.models import Workshop
 from WSS.models import WSS, Participant, UserProfile
 from WSS.payment import send_payment_request, verify
@@ -22,7 +22,7 @@ class WSSViewSet(viewsets.ViewSet):
     
     def list(self, request, year):
         wss = get_wss_object_or_404(year)
-        serializer = WSSSerializer(wss)
+        serializer = serializers.WSSSerializer(wss)
         return Response(serializer.data)
 
 
@@ -64,14 +64,14 @@ class BaseViewSet(viewsets.ViewSet, ABC):
 
 
 class WorkshopViewSet(BaseViewSet):
-    serializer = WorkshopSerializer
+    serializer = serializers.WorkshopSerializer
 
     def queryset_selector(self, request, wss):
         return wss.workshops
 
 
 class SeminarViewSet(BaseViewSet):
-    serializer = SeminarSerializer
+    serializer = serializers.SeminarSerializer
 
     def queryset_selector(self, request, wss):
         is_keynote = request.query_params.get("keynote", None)
@@ -81,14 +81,14 @@ class SeminarViewSet(BaseViewSet):
 
 
 class PosterSessionViewSet(BaseViewSet):
-    serializer = PosterSessionSerializer
+    serializer = serializers.PosterSessionSerializer
 
     def queryset_selector(self, request, wss):
         return wss.postersessions
 
 
 class SponsorshipViewSet(BaseViewSet):
-    serializer = SponsorshipSerializer
+    serializer = serializers.SponsorshipSerializer
 
     def queryset_selector(self, request, wss):
         is_main = request.query_params.get("main", None)
@@ -98,21 +98,21 @@ class SponsorshipViewSet(BaseViewSet):
 
 
 class ClipViewSet(BaseViewSet):
-    serializer = ClipSerializer
+    serializer = serializers.ClipSerializer
 
     def queryset_selector(self, request, wss):
         return wss.clips
 
 
 class HoldingTeamViewSet(BaseViewSet):
-    serializer = HoldingTeamSerializer
+    serializer = serializers.HoldingTeamSerializer
 
     def queryset_selector(self, request, wss):
         return wss.holding_teams
 
 
 class ImageViewSet(BaseViewSet):
-    serializer = ImageSerializer
+    serializer = serializers.ImageSerializer
 
     def queryset_selector(self, request, wss):
         return wss.images
@@ -128,7 +128,7 @@ class ErrorResponse(Response):
 class UserProfileViewSet(viewsets.ViewSet):
     authentication_classes = [SessionAuthentication, BasicAuthentication]
     permission_classes = [IsAuthenticated]
-    serializer = UserProfileSerializer
+    serializer = serializers.UserProfileSerializer
 
     def list(self, request):
         user_profile: UserProfile = request.user.profile
@@ -140,7 +140,7 @@ class UserProfileViewSet(viewsets.ViewSet):
         user_profile: UserProfile = request.user.profile
         user_data_parameter = request.data
 
-        fields = ['name', 'family', 'phone_number', 'age'
+        fields = ['first_name', 'last_name', 'phone_number', 'age'
                   'job', 'university', 'introduction_method',
                   'gender', 'city', 'country', 'field_of_interest',
                   'grade', 'is_student']
@@ -150,6 +150,7 @@ class UserProfileViewSet(viewsets.ViewSet):
                 setattr(user_profile, field, user_data_parameter[field])
         
         user_profile.save()
+        request.user.save()
         return Response(self.serializer(user_profile).data)
